Progetti_Oslo/IN3110-giorgich/assignment3/pyproject.toml
[build-system]
requires = [
    "setuptools>=61",
    # 4110 only:
    # "cython>=3",
]
build-backend = "setuptools.build_meta"

[project]
version = "1.0.0"
requires-python = ">=3.8"
license = {text = "MIT License"}
name = "instapy"
description = "Libray to apply filters to images, such as Black and White or sepia"
readme = "README.md"
dependencies = [
    "numpy",
    "numba",
    "pillow",
    "line-profiler"
]


[project.scripts]
instapy = "in3110_instapy.cli:main"
